requirements.txt
requests
pytest
Faker
allure-pytest
pytest-cov
selenium